Add MPD password functionality to POCOCM.
This is an initial patch that automatically sends MPD a connection password
if one has been provided to POCOCM->spawn(). Currently a normal mpd_error
event will be raised if the password is not correct, it might be worth making
this a fatal connect error at some point in the future.